requirements.txt
# Core ML Packages (Python 3.12 Compatible)
torch==2.2.0              
torchvision==0.17.0       
numpy==1.26.0             
opencv-python==4.9.0.80   

# Data Processing
albumentations==1.3.1      
pandas==2.1.4             
pyyaml==6.0.1             
scikit-learn==1.3.2       

# Visualization & Logging
matplotlib==3.8.0         
tensorboard==2.15.1       
pillow==10.1.0            

# Utilities
tqdm==4.66.1              
seaborn==0.13.0           
ipython==8.18.1           

# Optional Export Packages
onnx==1.15.0              
onnxruntime==1.17.0       

# Backward Compatibility Shims
typing-extensions==4.9.0  